+++ app/ReaderImage/CMakeLists.txt
@@ -10,11 +10,30 @@ execute_process(
 )
 
 set(INCLUDE_HEADERS "${CMAKE_CURRENT_SOURCE_DIR}/reader_img.hpp")
-set(SRC_FILES "${CMAKE_SOURCE_DIR}/src/read/reader_img.cpp")
+set(SRC_FILES "${CMAKE_CURRENT_SOURCE_DIR}/reader_img_s.cpp")
 add_library(ReadLib STATIC ${INCLUDE_HEADERS} ${SRC_FILES})
+
 
 set_target_properties(ReadLib PROPERTIES LINKER_LANGUAGE CXX)
 
 find_package( OpenCV REQUIRED PATHS "${CMAKE_SOURCE_DIR}/3rdparty/opencv/build" )
 include_directories( ${OpenCV_INCLUDE_DIRS} )
 target_link_libraries( ReadLib ${OpenCV_LIBS} )
+add_executable(Reader reader_img.cpp)
+target_link_libraries(Reader ReadLib)
+
+if (WIN32)
+add_custom_command(TARGET Reader POST_BUILD
+    COMMAND ${CMAKE_COMMAND} -E copy_directory
+        "${CMAKE_SOURCE_DIR}/3rdparty/opencv/build/bin/${CMAKE_BUILD_TYPE}"
+        "${CMAKE_BINARY_DIR}/app/ReaderImage/${CMAKE_BUILD_TYPE}/")
+endif()
+
+file(DOWNLOAD
+    "https://raw.githubusercontent.com/opencv/opencv/4.x/samples/data/lena.jpg"
+    "${CMAKE_CURRENT_BINARY_DIR}/image.jpg"
+    SHOW_PROGRESS
+    STATUS status_code
+    LOG log_file
+)
+add_definitions(-DIMAGE1_PATH="${CMAKE_CURRENT_BINARY_DIR}/image.jpg")
