+++ include/CMakeLists.txt
@@ -6,6 +6,3 @@ set(LAYERS_HEADERS "${layers_headers}" PARENT_SCOPE)
 
 file(GLOB_RECURSE perf_headers perf/*.h perf/*.hpp)
 set(PERF_HEADERS "${perf_headers}" PARENT_SCOPE)
-
-add_subdirectory(alexnet)
-add_subdirectory(read)
